refactor: Follow 패키지 follower -> follow 로 수정 및 dto에 유저 프사 추가
