+++ src/main/java/com/tablelog/tablelogback/domain/follow/dto/FollowUserDto.java
@@ -1,8 +1,9 @@
-package com.tablelog.tablelogback.domain.follower.dto;
+package com.tablelog.tablelogback.domain.follow.dto;
 
 public record FollowUserDto(
         Long userId,
         String nickname,
+        String profileImgUrl,
         Boolean isFollowed
 ) {
 }
