+++ src/main/java/com/tablelog/tablelogback/domain/follow/exception/AlreadyExistsFollowingException.java
@@ -1,4 +1,4 @@
-package com.tablelog.tablelogback.domain.follower.exception;
+package com.tablelog.tablelogback.domain.follow.exception;
 
 import com.tablelog.tablelogback.global.exception.CustomException;
 import com.tablelog.tablelogback.global.exception.ErrorCode;
